+++ .gitignore
@@ -8,6 +8,7 @@
 .idea/**/usage.statistics.xml
 .idea/**/dictionaries
 .idea/**/shelf
+application/src/main/resources/privates
 
 # AWS User-specific
 .idea/**/aws.xml
