+++ application/src/main/java/core/application/config/PrivateConfig.java
@@ -5,4 +5,10 @@ import org.springframework.context.annotation.PropertySource;
 
 @Configuration
 @PropertySource("classpath:privates/private.properties")
-public class PrivateConfig {}
+public class PrivateConfig {
+    /*  privates/private.properties
+      ${database.url}       :   DB url
+      ${database.username}  :   DB 접속 ID
+      ${database.password}  :   DB 접속 PW
+     */
+}
