--- application/src/main/java/core/application/models/entities/UserRole.java
@@ -1,5 +1,0 @@
-package core.application.models.entities;
-
-public enum UserRole {
-    ADMIN, USER
-}
