+++ application/src/main/java/core/application/users/models/entities/DibEntity.java
@@ -1,4 +1,4 @@
-package core.application.models.entities;
+package core.application.users.models.entities;
 
 import lombok.Data;
 
@@ -8,5 +8,5 @@ import java.util.UUID;
 public class DibEntity {
     private Long   DibId;
     private UUID   userId;
-    private String movieId;     // 영화 API 에 따라 달라질 수 있음.
+    private String movieId;
 }
