+++ application/src/main/java/core/application/users/models/entities/UserEntity.java
@@ -1,4 +1,4 @@
-package core.application.models.entities;
+package core.application.users.models.entities;
 
 import lombok.Data;
