+++ application/src/main/java/core/application/users/models/entities/UserRole.java
@@ -1,0 +1,5 @@
+package core.application.users.models.entities;
+
+public enum UserRole {
+    ADMIN, USER
+}
